Consolidate ExecutableCreator implementation

Copilot-Session: a447f900-bb02-4165-ae7b-9872040bc4fe
