+++ src/Aspire.Hosting/DistributedApplicationBuilder.cs
@@ -545,11 +545,8 @@ public class DistributedApplicationBuilder : IDistributedApplicationBuilder
             // DCP stuff
             _innerBuilder.Services.AddSingleton<DcpAppResourceStore>();
             _innerBuilder.Services.AddSingleton<ProxylessEndpointPortAllocator>();
-            _innerBuilder.Services.AddSingleton<ExecutableResourcePreparer>();
             _innerBuilder.Services.AddSingleton<ExecutableConfigurationResolver>();
             _innerBuilder.Services.AddSingleton<ExecutableLaunchPolicy>();
-            _innerBuilder.Services.AddSingleton<ExecutableLaunchPlanResolver>();
-            _innerBuilder.Services.AddSingleton<DcpExecutableRenderer>();
             _innerBuilder.Services.AddSingleton<ExecutableCreator>();
             _innerBuilder.Services.AddSingleton<ContainerCreator>();
             _innerBuilder.Services.AddSingleton<DcpExecutor>();
